cripto/bot_trader_binance/requirements.txt
python-binance==1.0.29
python-dotenv==1.1.1
pandas==2.3.1